+++ apps/gizual-app/src/primitives/canvas/components/file-canvas.tsx
@@ -21,6 +21,8 @@ type MasonryCanvasProps = {
  */
 export const MasonryCanvas = observer(({ vm, wrapper }: MasonryCanvasProps) => {
   const mainController = useMainController();
+  const settingsController = mainController.settingsController;
+
   const blocks = useContext(CanvasContext).useBlocks();
 
   console.log({ blocks, wrapper, cw: vm.canvasWidth });
@@ -34,10 +36,10 @@ export const MasonryCanvas = observer(({ vm, wrapper }: MasonryCanvasProps) => {
       )}
 
       <MasonryGrid
-        width={vm.canvasWidth}
         childInfo={blocks.map((f) => {
           return { id: f.id, height: f.height + 26 };
         })}
+        numColumns={settingsController.settings.visualizationSettings.canvas.masonryColumns.value}
       >
         {blocks.map((block, index) => {
           if (!wrapper) return <React.Fragment key={index}></React.Fragment>;